profiler/requirements.txt
asyncio==3.4.3
asyncpg==0.24.0
python-dotenv==0.19.1
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
